--- site/cookbook_gallery.yaml
@@ -1,13 +1,0 @@
-- title: Radar Cookbook
-  url: https://projectpythiatutorials.github.io/radar-cookbook/landing-page.html
-  description: |
-    This Project Pythia Cookbook covers the basics of working with weather radar data in Python.
-  authors:
-  - name: Max Grover
-  thumbnail: _static/images/thumbnails/arm_logo.png
-  tags:
-    domains:
-    - radar
-    packages:
-    - py-art
-
